[#36] Refactor: 패키지 구조 정리 및 이름 통일
